change the type of variable for calculateCredit and Math.round(pay)

+++ src/main/java/ru/itpark/CreditService.java
@@ -3,12 +3,10 @@ package ru.itpark;
 
 
     public class CreditService {
-        public double calculateCredit(double creditSum, double creditPeriod, double creditRate) {
-
+        public long calculateCredit(double creditSum, double creditPeriod, double creditRate) {
 
 
             double monthSum = creditRate/100/12;
-
 
 
             double pay = creditSum * (monthSum + (monthSum /(Math.pow((1 + monthSum), creditPeriod) - 1)));
